profile \ profiles controller merging
